Splitted directoryd.proto and segregated relevant files
