+++ orc8r/cloud/go/services/directoryd/directoryd/main.go
@@ -38,11 +38,12 @@ import (
 	"magma/orc8r/cloud/go/orc8r"
 	"magma/orc8r/cloud/go/service"
 	"magma/orc8r/cloud/go/services/directoryd"
+	"magma/orc8r/cloud/go/services/directoryd/protos"
 	servicers "magma/orc8r/cloud/go/services/directoryd/servicers/protected"
 	dstorage "magma/orc8r/cloud/go/services/directoryd/storage"
 	"magma/orc8r/cloud/go/sqorc"
 	"magma/orc8r/cloud/go/storage"
-	"magma/orc8r/lib/go/protos"
+	lib_protos "magma/orc8r/lib/go/protos"
 )
 
 func main() {
@@ -73,7 +74,7 @@ func main() {
 		glog.Fatalf("Error creating initializing directory servicer: %s", err)
 	}
 	protos.RegisterDirectoryLookupServer(srv.GrpcServer, servicer)
-	protos.RegisterGatewayDirectoryServiceServer(srv.GrpcServer, servicers.NewDirectoryUpdateServicer())
+	lib_protos.RegisterGatewayDirectoryServiceServer(srv.GrpcServer, servicers.NewDirectoryUpdateServicer())
 
 	// Run service
 	err = srv.Run()
