+++ orc8r/cloud/go/services/directoryd/protos/directoryd_helper.go
