+++ orc8r/cloud/go/services/directoryd/servicers/protected/servicer_test.go
@@ -20,11 +20,11 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"magma/orc8r/cloud/go/blobstore"
+	"magma/orc8r/cloud/go/services/directoryd/protos"
 	servicers "magma/orc8r/cloud/go/services/directoryd/servicers/protected"
 	"magma/orc8r/cloud/go/services/directoryd/storage"
 	stateTestInit "magma/orc8r/cloud/go/services/state/test_init"
 	"magma/orc8r/cloud/go/sqorc"
-	"magma/orc8r/lib/go/protos"
 )
 
 const (
