+++ orc8r/cloud/go/services/directoryd/test_init/test_service_init.go
@@ -21,11 +21,12 @@ import (
 	"magma/orc8r/cloud/go/blobstore"
 	"magma/orc8r/cloud/go/orc8r"
 	"magma/orc8r/cloud/go/services/directoryd"
+	"magma/orc8r/cloud/go/services/directoryd/protos"
 	servicers "magma/orc8r/cloud/go/services/directoryd/servicers/protected"
 	"magma/orc8r/cloud/go/services/directoryd/storage"
 	"magma/orc8r/cloud/go/sqorc"
 	"magma/orc8r/cloud/go/test_utils"
-	"magma/orc8r/lib/go/protos"
+	lib_protos "magma/orc8r/lib/go/protos"
 )
 
 func StartTestService(t *testing.T) {
@@ -44,7 +45,7 @@ func StartTestService(t *testing.T) {
 	directoryServicer, err := servicers.NewDirectoryLookupServicer(store)
 	assert.NoError(t, err)
 	protos.RegisterDirectoryLookupServer(srv.GrpcServer, directoryServicer)
-	protos.RegisterGatewayDirectoryServiceServer(srv.GrpcServer, servicers.NewDirectoryUpdateServicer())
+	lib_protos.RegisterGatewayDirectoryServiceServer(srv.GrpcServer, servicers.NewDirectoryUpdateServicer())
 
 	// Run service
 	go srv.RunTest(lis)
